updated sending, added username, limited characters
Changed the way data is sent, first it echoes to the rx box, then passes
that string to the comport writer, and eliminates the repetitiveness of
including the carriage return. Also added a username function so that
users can id who's talking; limited to 8 characters. Limited the send
text to 128 characters. All the set functions will echo to the rx box.