added crypto
Added cryptography option to sending messages. Uses an obscured password
(8 char min), where text being sent is encoded and transmitted, as well
as decoded when received. Had to move the senders username into the
encryptor so that it wouldn't mess up the decyrptor. Now you'll get a
message telling you it's been decoded and who it's from. Similarly, also
implemented some AT catching so that words like "that?" wont get caught
as AT commands and instead are encoded and sent.